--- notes v2.txt
@@ -1,31 +1,0 @@
-COMMIT NOTES
-previous commit: shrank header and footer, changed font sizes on Tpot42, introduced random ordering to Tpot 42
-current commit: 
-
-TO DO
-    MENU WITH SUBMENUS or NEW ORG so you can get places easily
-        MEDIA
-            illustration
-            animation
-        PROJECTS
-            ingroup map
-            tpot for two
-            goblin friender
-        SOCIAL
-            twitter <-- link to code twitter
-            instagram
-            youtube
-        SUPPORT
-            patreon
-            ko-fi
-            redbubble
-
-    TPOT 42
-        add an age filter
-        add a second location filter
-        scrolling (or pages) so you can see there are other directories
-        TANUKI suggestion re: CSS
-            https://twitter.com/goblincodes/status/1563274182439231489?s=20&t=Xu5LFktJAjwNbkvbXRqfhg
-        black borders around menu names onclick
-
-check if notesTEST.txt still exists after you deploy
